@@ -4,13 +4,13 @@ import { MatCheckbox } from '@angular/material';
 
 @Component({
   selector: 'mat-multiselect',
-  templateUrl: './multiselect.component.html',
-  styleUrls: ['./multiselect.component.scss']
+  templateUrl: './mat-multiselect.component.html',
+  styleUrls: ['./mat-multiselect.component.scss']
 })
 /**
  * Multiselect Component which control a patch of screen containing multiselect dropdown control
  */
-export class MultiselectComponent implements OnInit {
+export class MatMultiselectComponent implements OnInit {
   /**
    * Input Data to be passed from parent in the following format
    * { placeHolder: 'Locations', enableSelectClear: true, enableSearch: true }
@@ -21,17 +21,17 @@ export class MultiselectComponent implements OnInit {
    * Input Options to be passed from parent in the following format array
    * [ {name: 'ind', value: 'India'} ]
    */
-  @Input('options') primalOptions: any[];
+  @Input() options: any[];
 
   /**
    * Output to parent for selected options
    */
-  @Output() changeSelection = new EventEmitter < any[] > ();
+  @Output() changeSelection = new EventEmitter<any[]>();
 
   /**
    * Child mat checkbox
    */
-  @ViewChild(MatCheckbox) matCheckbox;
+  @ViewChild(MatCheckbox, { static: false }) matCheckbox;
 
   /**
    * form control for multiselect dropdown
@@ -56,13 +56,13 @@ export class MultiselectComponent implements OnInit {
   /**
    * Constructor
    */
-  constructor() {}
+  constructor() { }
 
   /**
    * Angular lifecycle event : run once after ngOnChanges
    */
   ngOnInit() {
-      this.emitValue();
+    this.emitValue();
   }
 
   /**
@@ -70,42 +70,42 @@ export class MultiselectComponent implements OnInit {
    * @param option selected option in multiselect dropdown
    */
   optionClick(option) {
-      const index = this.primalOptions.indexOf(option);
-      this.primalOptions[index].checked = !this.primalOptions[index].checked;
-      this.emitValue();
+    const index = this.options.indexOf(option);
+    this.options[index].checked = !this.options[index].checked;
+    this.emitValue();
   }
 
   /**
    * Event to emit selected values to parent throgh changeSelection event emmiter
    */
   emitValue() {
-      this.selectedOptions = this.primalOptions.filter(option => option.checked);
-      this.multiSelect.patchValue(this.selectedOptions);
-      this.evaluateStates();
-      this.changeSelection.emit(this.selectedOptions);
+    this.selectedOptions = this.options.filter(option => option.checked);
+    this.multiSelect.patchValue(this.selectedOptions);
+    this.evaluateStates();
+    this.changeSelection.emit(this.selectedOptions);
   }
 
   /**
    * Event for changing the state of select all checkbox according to option's state
-   * @param value 
+   * @param value checked value
    */
-  changeState(value) {
-      this.matSearch = '';
-      this.primalOptions.forEach(option => option.checked = value);
-      this.emitValue();
+  changeState(value: boolean) {
+    this.matSearch = '';
+    this.options.forEach(option => option.checked = value);
+    this.emitValue();
   }
 
   /**
    * Event for evaluating the state of select all checkbox according to option's state
    */
   evaluateStates() {
-      this.indeterminate = false;
-      if (this.selectedOptions.length == 0 && this.matCheckbox != undefined) {
-          this.matCheckbox.checked = false;
-      } else if (this.primalOptions.length == this.selectedOptions.length && this.matCheckbox != undefined) {
-          this.matCheckbox.checked = true;
-      } else {
-          this.indeterminate = true;
-      }
+    this.indeterminate = false;
+    if (this.selectedOptions.length === 0 && this.matCheckbox !== undefined) {
+      this.matCheckbox.checked = false;
+    } else if (this.options.length === this.selectedOptions.length && this.matCheckbox !== undefined) {
+      this.matCheckbox.checked = true;
+    } else {
+      this.indeterminate = true;
+    }
   }
 }
