@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'ms-root',
+  selector: 'mat-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
@@ -10,15 +10,16 @@ export class AppComponent {
     placeHolder: 'Locations',
     enableSelectClear: true,
     enableSearch: true
-  }
+  };
   options = [
-    {name: 'ind', value: 'India'}, 
-    {name: 'nep', value: 'Nepal', checked: true},
-    {name: 'bon', value: 'Bondhar'},
-    {name: 'Kat', value: 'Kathmandu', checked: true},
-    {name: 'ner', value: 'Nerli'},
-    {name: 'bom', value: 'Bombay'},];
-  changeSelection(options){
-    console.log(options);
+    { name: 'ind', value: 'India' },
+    { name: 'nep', value: 'Nepal', checked: true },
+    { name: 'bon', value: 'Bondhar' },
+    { name: 'Kat', value: 'Kathmandu', checked: true },
+    { name: 'ner', value: 'Nerli' },
+    { name: 'bom', value: 'Bombay' }
+  ];
+  changeSelection(options) {
+    console.table(options);
   }
 }
